Major Update of the Add-In Version 3
This is a major rewrite of the code to convert it from Javascript to a Python based add-in.  Fusion 360 has ended support for Javascript therefore the change.

Additionally, the design and implementation has been changed and is hopefully is bit easier and more interactive to use.

There are still issues and todos but those are coming.  Overall, the functionality should be working.